[T6811]bad_connector_woocommerce:Added exporter.
