+++ bad_connector_woocommerce/models/product/__init__.py
@@ -1,2 +1,3 @@
 from . import common
 from . import importer
+from . import exporter
