+++ bad_connector_woocommerce/models/product/importer.py
@@ -361,17 +361,3 @@ class WooProductProductImporter(Component):
         image_importer = self.component(usage="product.image.importer")
         image_importer.run(self.external_id, binding, image_record)
         return result
-
-
-class ProductInventoryExporter(Component):
-    _name = "woo.product.product.exporter"
-    _inherit = "woo.exporter"
-    _apply_on = ["woo.product.product"]
-    _usage = "product.inventory.exporter"
-
-    def run(self, binding, record=None, *args, **kwargs):
-        """Export the product inventory to WooCommerce"""
-        external_id = self.binder.to_external(binding)
-        data = {"stock_quantity": binding.woo_product_qty}
-        if binding.backend_id.update_stock_inventory and binding.stock_management:
-            self.backend_adapter.write(external_id, data)
