+++ bad_connector_woocommerce/models/woo_binding/common.py
@@ -55,3 +55,10 @@ class WooBinding(models.AbstractModel):
         with backend.work_on(self._name) as work:
             exporter = work.component(usage="record.exporter")
             return exporter.run(self, record, fields)
+
+    def export_inventory(self, fields=None):
+        """Export the Quantity of a Product."""
+        self.ensure_one()
+        with self.backend_id.work_on(self._name) as work:
+            exporter = work.component(usage="product.inventory.exporter")
+            return exporter.run(self, fields)
